rename Payload to MetricInput
